% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ckde_dates.R
\name{calc_ckde_dates}
\alias{calc_ckde_dates}
\title{Calculate Composite Kernel Density Estimates of Radiocarbon Dates}
\usage{
calc_ckde_dates(data, interval = 0.95, timeRange_max, timeRange_min)
}
\arguments{
\item{data}{A \code{data.frame} containing radiocarbon dates.}

\item{interval}{A \code{numeric} value for the quantile interval. Default is 0.95.}

\item{timeRange_max}{A \code{numeric} value for the maximum time range in BP.}

\item{timeRange_min}{A \code{numeric} value for the minimum time range in BP.}
}
\value{
A tibble that has the mean density estimate, the high and low values for the quantile interval, and a BC/AD date.
}
\description{
\code{calc_ckde_dates()} calibrates the uncalibrated radiocarbon dates, then samples the dates with simulation and bootstrap,
then returns the mean density estimate, the high and low values for the quantile interval, and a BC/AD date column as a dataframe.
}
