% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSPD.R
\name{makeSPD}
\alias{makeSPD}
\title{Calibrate Dates and Calculate SPDs}
\usage{
makeSPD(data)
}
\arguments{
\item{data}{A data.frame containing radiocarbon dates.}
}
\value{
A tibble that has a \code{CalGrid} class object containing the summed probability associated to each calendar year between timeRange\link{1} and timeRange\link{2}.
}
\description{
\code{makeSPD()} calibrates radiocarbon dates and calculates a summed probability distributions (SPD) of radiocarbon dates.
}
