% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRISM_gs_gdd}
\alias{PRISM_gs_gdd}
\title{Growing Season Accumulated Growing Degree Days (째C) from PRISM}
\format{
An object of class \code{RasterLayer}.
}
\source{
\url{http://www.prism.oregonstate.edu/}
}
\usage{
PRISM_gs_gdd
}
\description{
A \code{raster} containing accumulated growing degree days (째C) for the growing season (May through September) with a base temperature of 10째C and maximum temperature of 30째C.
GDD was calculated from tmin and max 30 seconds (~1 \ifelse{html}{\out{km<sup>2</sup>}}{\eqn{km^2}}).
}
\keyword{datasets}
