% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_dag_arrows.R
\name{shorten_dag_arrows}
\alias{shorten_dag_arrows}
\title{Adjust Length of Dag Arrows}
\usage{
shorten_dag_arrows(tidy_dag, proportion)
}
\arguments{
\item{tidy_dag}{A \code{tidy_dagitty} object containing a tidy dag.}
}
\value{
A tidy_dagitty with the adjusted dag arrows.
}
\description{
\code{shorten_dag_arrows()} scales down dag arrows from both ends according to a proportion.
}
