% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPD_county}
\alias{SPD_county}
\title{Summed Probability Distributions for Four Corners Counties}
\format{
An object of class \code{tbl_df}.
\describe{
\item{\code{NAME}}{County of the site from which the sample has been recovered}
\item{\code{REGION}}{State of the site from which the sample has been recovered}
\item{\code{n}}{Sample size of radiocarbon dates for each county}
\item{\code{data}}{Radiocarbon samples in each county as a nested \code{tibble}}
\item{\code{spds}}{SPDs for each county as an object of class \code{CalGrid}}
\item{\code{county_st}}{County and state of the site from which the sample has been recovered}
}
}
\usage{
SPD_county
}
\description{
A \code{tibble} containing SPDs for counties in Arizona, Colorado, New Mexico, and Utah. Also, contains the number of samples and original data used to make the SPD.
}
\keyword{datasets}
