% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw_rc}
\alias{sw_rc}
\title{Radiocarbon Data for Four Corners}
\format{
An object of class \code{tbl_df} with the following variables:
\describe{
\item{\code{LabID}}{Lab ID of each radiocarbon date}
\item{\code{SiteID}}{ID of the site from which the sample has been recovered}
\item{\code{SiteName}}{Name of the site from which the sample has been recovered}
\item{\code{county_st}}{County and state of the site from which the sample has been recovered}
\item{\code{NAME}}{County of the site from which the sample has been recovered}
\item{\code{REGION}}{State of the site from which the sample has been recovered}
\item{\code{Taxa}}{Material of the dated sample}
\item{\code{Age}}{Radiocarbon age in 14C years BP}
\item{\code{Error}}{Radiocarbon age error}
\item{\code{calib}}{Calibration curve to be used to calibrate the Age}
}
}
\source{
\url{https://github.com/people3k/p3k14c}

\url{https://github.com/Archaeo-Programmer/cropDiffusionR/blob/main/vignettes/tables}
}
\usage{
sw_rc
}
\description{
A \code{data.frame} containing radiocarbon data for the Four Corners. \code{p3k14c} data was combined with \code{cropDiffusionR}.
}
\keyword{datasets}
