% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swus_NED}
\alias{swus_NED}
\title{National Elevation Dataset for Southwestern United States}
\format{
An object of class \code{RasterLayer}.
}
\source{
\url{https://ned.usgs.gov/}
}
\usage{
swus_NED
}
\description{
A \code{raster} containing elevation data for the Southwestern United States.
}
\keyword{datasets}
