% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPD_county}
\alias{SPD_county}
\title{Summed Probability Distributions for Four Corners Counties.}
\format{
An object of class \code{tbl_df}.
}
\usage{
SPD_county
}
\description{
A \code{tibble} containing SPDs for counties in Arizona, Colorado, New Mexico, and Utah. Also, contains the number of samples and original data used to make the SPD.
}
\keyword{datasets}
