% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SWUS_counties}
\alias{SWUS_counties}
\title{Four Corners County Shapefiles.}
\format{
An object of class \code{sfc_POLYGON}.
}
\source{
\url{https://github.com/adeckmyn/maps}
}
\usage{
SWUS_counties
}
\description{
A \code{sf} \code{data.frame} containing county shapefiles for Arizona, Colorado, New Mexico, and Utah.
}
\keyword{datasets}
